api/requirements.txt
Django==3.1
djangorestframework==3.11.1
pygments
django-cors-headers==3.10.1
numpy
opencv-python
tensorflow==1.14.0
Cython